+++ packages/ui/src/recoil/settings/atom.ts
@@ -9,8 +9,6 @@ const initialState: AtomState = {
   dateFormat: 'locale',
   timeFormat: '12-hour',
   txListFormat: 'compact',
-  filterMsgTypes: '{}',
-  openDialog: false,
 };
 
 export const atomState = atom<AtomState>({
