+++ packages/ui/src/recoil/settings/index.ts
@@ -5,14 +5,10 @@ export {
   readTimeFormat,
   readTheme,
   readTx,
-  readFilterMsgTypes,
   writeDate,
   writeTimeFormat,
   writeTheme,
   writeTx,
-  writeFilterMsgTypes,
-  writeOpenDialog,
-  readOpenDialog,
 } from '@/recoil/settings/selectors';
 export type { AtomState, Date, Theme, Tx, TimeFormat } from '@/recoil/settings/types';
 export {
