+++ packages/ui/src/recoil/settings/selectors.ts
@@ -115,49 +115,3 @@ export const readTx = selector({
   key: 'settings.read.tx',
   get: getTx,
 });
-
-const getFilterMsgTypes: ReadOnlySelectorOptions<string>['get'] = ({ get }) => {
-  const state = get(atomState);
-  return state.filterMsgTypes;
-};
-
-export const writeFilterMsgTypes = selector({
-  key: 'settings.write.filterMsgTypes',
-  get: getFilterMsgTypes,
-  set: ({ get, set }, newMsgFilter) => {
-    if (newMsgFilter instanceof DefaultValue) return;
-    const prevState = get(atomState);
-    const newState = mergeStateChange(prevState, {
-      filterMsgTypes: newMsgFilter,
-    });
-    set(atomState, newState);
-  },
-});
-
-export const readFilterMsgTypes = selector({
-  key: 'settings.read.filterMsgTypes',
-  get: getFilterMsgTypes,
-});
-
-const getOpenDialog: ReadOnlySelectorOptions<boolean>['get'] = ({ get }) => {
-  const state = get(atomState);
-  return state.openDialog;
-};
-
-export const writeOpenDialog = selector({
-  key: 'settings.write.openDialog',
-  get: getOpenDialog,
-  set: ({ get, set }, newOpenDialogValue) => {
-    if (newOpenDialogValue instanceof DefaultValue) return;
-    const prevState = get(atomState);
-    const newState = mergeStateChange(prevState, {
-      openDialog: newOpenDialogValue,
-    });
-    set(atomState, newState);
-  },
-});
-
-export const readOpenDialog = selector({
-  key: 'settings.read.openDialog',
-  get: getOpenDialog,
-});
