+++ packages/ui/src/recoil/transactions_filter/atom.ts
@@ -1,0 +1,12 @@
+import { atom } from 'recoil';
+import type { AtomState } from '@/recoil/transactions_filter/types';
+
+const initialState: AtomState = {
+  filterMsgTypes: '{}',
+  openDialog: false,
+};
+
+export const atomState = atom<AtomState>({
+  key: 'txsFilter',
+  default: initialState,
+});
