+++ packages/ui/src/recoil/transactions_filter/hooks.ts
@@ -1,0 +1,20 @@
+import { useEffect } from 'react';
+import { useRecoilState } from 'recoil';
+import { atomState } from '@/recoil/transactions_filter/atom';
+import type { AtomState } from '@/recoil/transactions_filter/types';
+
+const isClient = typeof window === 'object';
+
+export const useTxsFilterRecoil = () => {
+  const [txsFilter, setTxsFilter] = useRecoilState(atomState);
+
+  useEffect(() => {
+    if (isClient) {
+      const initTxsFilter: AtomState = {
+        filterMsgTypes: '{}',
+        openDialog: false,
+      };
+      setTxsFilter(initTxsFilter);
+    }
+  }, [setTxsFilter, txsFilter.filterMsgTypes, txsFilter.openDialog]);
+};
