+++ packages/ui/src/recoil/transactions_filter/index.ts
@@ -1,0 +1,9 @@
+export { atomState } from '@/recoil/transactions_filter/atom';
+export { useTxsFilterRecoil } from '@/recoil/transactions_filter/hooks';
+export {
+  readFilterMsgTypes,
+  writeFilterMsgTypes,
+  writeOpenDialog,
+  readOpenDialog,
+} from '@/recoil/transactions_filter/selectors';
+export type { AtomState } from '@/recoil/transactions_filter/types';
