+++ packages/ui/src/recoil/transactions_filter/types.ts
@@ -1,0 +1,4 @@
+export interface AtomState {
+  filterMsgTypes: string;
+  openDialog: boolean;
+}
