+++ packages/ui/src/screens/account_details/components/transactions/hooks.ts
@@ -9,7 +9,7 @@ import {
 import type { TransactionState } from '@/screens/account_details/components/transactions/types';
 import { convertMsgType } from '@/utils/convert_msg_type';
 import { useRecoilValue } from 'recoil';
-import { readFilterMsgTypes } from '@/recoil/settings';
+import { readFilterMsgTypes } from '@/recoil/transactions_filter';
 
 const LIMIT = 50;
