+++ packages/ui/src/screens/transactions/index.tsx
@@ -6,7 +6,8 @@ import Layout from '@/components/layout';
 import LoadAndExist from '@/components/load_and_exist';
 import TransactionsList from '@/components/transactions_list';
 import TransactionsListDetails from '@/components/transactions_list_details';
-import { readTx, writeFilterMsgTypes } from '@/recoil/settings';
+import { readTx } from '@/recoil/settings';
+import { writeFilterMsgTypes } from '@/recoil/transactions_filter';
 import { useTransactions } from '@/screens/transactions/hooks';
 import useStyles from '@/screens/transactions/styles';
 import { useEffect } from 'react';
